The io.mockk.gh package tests in the mockk-jvm module are moved and renamed to better describe which features are being tested.
This commit is part of #618.
