--- mockk/jvm/src/test/java/io/mockk/gh/TestBarIssue119.java
@@ -1,4 +1,0 @@
-package io.mockk.gh;
-
-class TestBarIssue119 {
-}
