+++ mockk/jvm/src/test/java/io/mockk/it/PackagePrivateCls.java
@@ -1,0 +1,4 @@
+package io.mockk.it;
+
+class PackagePrivateCls {
+}
