+++ mockk/jvm/src/test/kotlin/io/mockk/it/SafeCallSuspendingTest.kt
@@ -1,12 +1,16 @@
-@file:Suppress("VARIABLE_WITH_REDUNDANT_INITIALIZER", "UNNECESSARY_SAFE_CALL")
+@file:Suppress("VARIABLE_WITH_REDUNDANT_INITIALIZER", "UNNECESSARY_SAFE_CALL", "RedundantSuspendModifier")
 
-package io.mockk.gh
+package io.mockk.it
 
 import io.mockk.mockk
 import kotlinx.coroutines.runBlocking
 import org.junit.Test
 
-class Issue174Test {
+/**
+ * Safe call operator throws a class cast exception on relaxed mock of suspending function.
+ * Verifies issue #174.
+ */
+class SafeCallSuspendingTest {
     class MyClass {
         suspend fun doSomething() {
 
@@ -14,7 +18,7 @@ class Issue174Test {
     }
 
     @Test
-    fun test() {
+    fun testSafeCall() {
         runBlocking {
             var myPossiblyNullInstance: MyClass? = null
             myPossiblyNullInstance = mockk(relaxed = true)
